+++ R/ceplane_plot_graph.R
@@ -143,26 +143,28 @@ ceplane_plot_ggplot.bcea <- function(he,
   }
   
   # single long format for ggplot data
-  delta_c <- he$delta_c |>
-    as_tibble() |>
-    mutate(sim = row_number()) |>
-    pivot_longer(
-      cols = -"sim",
-      names_to = "comparison",
-      values_to = "delta_c"
-    )
+  # delta_c <- he$delta_c |>
+  #   as_tibble() |>
+  #   mutate(sim = row_number()) |>
+  #   pivot_longer(
+  #     cols = -"sim",
+  #     names_to = "comparison",
+  #     values_to = "delta_c"
+  #   )
+  # 
+  # delta_e <- he$delta_e |>
+  #   as_tibble() |>
+  #   mutate(sim = row_number()) |>
+  #   pivot_longer(
+  #     cols = -"sim",
+  #     names_to = "comparison",
+  #     values_to = "delta_e"
+  #   )
+  # 
+  # delta_ce <- left_join(delta_c, delta_e, by = c("sim", "comparison")) |>
+  #   mutate(comparison=factor(comparison))
   
-  delta_e <- he$delta_e |>
-    as_tibble() |>
-    mutate(sim = row_number()) |>
-    pivot_longer(
-      cols = -"sim",
-      names_to = "comparison",
-      values_to = "delta_e"
-    )
-  
-  delta_ce <- left_join(delta_c, delta_e, by = c("sim", "comparison")) |>
-    mutate(comparison=factor(comparison))
+  delta_ce <- prep_delta_ce(he)
   
   plot_params <-
     ceplane_ggplot_params(he, pos_legend = pos_legend, graph_params = graph_params, ...)
@@ -241,26 +243,28 @@ ceplane_plot_plotly.bcea <- function(he,
   wtp <- graph_params$wtp_value
   
   # single long format for plotting data
-  delta_c <- he$delta_c |>
-    as_tibble() |>
-    mutate(sim = row_number()) |>
-    pivot_longer(
-      cols = -"sim",
-      names_to = "comparison",
-      values_to = "delta_c"
-    )
+  # delta_c <- he$delta_c |>
+  #   as_tibble() |>
+  #   mutate(sim = row_number()) |>
+  #   pivot_longer(
+  #     cols = -"sim",
+  #     names_to = "comparison",
+  #     values_to = "delta_c"
+  #   )
+  # 
+  # delta_e <- he$delta_e |>
+  #   as_tibble() |>
+  #   mutate(sim = row_number()) |>
+  #   pivot_longer(
+  #     cols = -"sim",
+  #     names_to = "comparison",
+  #     values_to = "delta_e"
+  #   )
+  # 
+  # delta_ce <- left_join(delta_c, delta_e, by = c("sim", "comparison")) |>
+  #   mutate(comparison=factor(comparison))
   
-  delta_e <- he$delta_e |>
-    as_tibble() |>
-    mutate(sim = row_number()) |>
-    pivot_longer(
-      cols = -"sim",
-      names_to = "comparison",
-      values_to = "delta_e"
-    )
-  
-  delta_ce <- left_join(delta_c, delta_e, by = c("sim", "comparison")) |>
-    mutate(comparison=factor(comparison))
+  delta_ce <- prep_delta_ce(he)
   
   # Adjust graph parameters
   if (length(graph_params$point$color) != length(comp_label)) {