+++ R/plot.mixedAn.R
@@ -23,7 +23,8 @@
 #' corner inside the plot area.
 #' @param graph A string used to select the graphical engine to use for
 #' plotting. Should (partial-)match the two options `"base"` or
-#' `"ggplot2"`. Default value is `"base"`.
+#' `"ggplot2"`. Default value is `"base"`. The `"plotly"` option is not 
+#' implemented for this particular graph.
 #' @param ...  Arguments to be passed to methods, such as graphical parameters
 #' (see [par()]).
 #' @return \item{evi}{ A ggplot object containing the plot. Returned only if
@@ -34,10 +35,32 @@
 #' @author Gianluca Baio, Andrea Berardi
 #' @usage plot.mixedAn(x, y.limits=NULL, pos=c(0,1), graph=c("base","ggplot2"),...)
 #'
+#' \dontrun{
+#' # Load the processed results of the MCMC simulation model
+#' data(Vaccine)
+#' # 
+#' # Runs the health economic evaluation using BCEA
+#' m <- bcea(e=eff, c=cost,    # defines the variables of 
+#'                             #  effectiveness and cost
+#'       ref=2,                # selects the 2nd row of (e,c) 
+#'                             #  as containing the reference intervention
+#'       interventions=treats, # defines the labels to be associated 
+#'                             #  with each intervention
+#'       Kmax=50000,           # maximum value possible for the willingness 
+#'                             #  to pay threshold; implies that k is chosen 
+#'                             #  in a grid from the interval (0,Kmax)
+#'       plot=FALSE            # inhibits graphical output
+#' )
+#' # Runs the mixed Analysis and plots the results
+#' x=`mixedAn<-`(m)            # Creates the 'mixedAn' object
+#' plot(x)                     # And plots the results
+#' }
+#' 
 #' @export
 #' 
 plot.mixedAn <- function(x, y.limits = NULL, pos = c(0,1),
                          graph = c("base","ggplot2"), ...) {
   .Deprecated(new = "evi.plot")
+  evi.plot(`mixedAn<-`(x),y.limits=NULL, pos=c(0,1),graph=graph,...)
 }